Find the Missing Number.cpp
You are given an array arr containing n-1 distinct integers. The array consists of integers taken from the range 1 to n, meaning one integer is missing from this sequence. Your task is to find the missing integer.

Input :
An integer array arr of size n-1 where the elements are distinct and taken from the range 1 to n.
Example : arr = [1, 2, 4, 5]

Output :
Return the missing integer from the array.
Example: Missing number: 3

+++ Day2.cpp
@@ -8,7 +8,7 @@ int main() {
     cout<<"Enter n:";
     cin>>n;
     
-    int arr[n-1];
+    long long arr[n-1];
     cout<<"Enter a "<<n-1<<" Element in array(1 to n):"<<endl;
     
     for(int i=0; i<n-1; i++) {
@@ -16,8 +16,8 @@ int main() {
     }
     
     //*** MAIN CODE ****
-    int expected_sum = (n*(n+1))/2;
-    int curr_sum = 0;
+    long long expected_sum = (n*(n+1))/2;
+    long long curr_sum = 0;
     
     for(int &val: arr) {
         curr_sum += val;
